backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2025.2
urllib3==2.3.0
